Fix for http://dev.exiv2.org/issues/1213
Two changes: add $(ProjectDir)../../src to the include path
change $(SolutionDir)../ to $(ProjectDir)../../ in order so that individual projects may be used in solutions other than those in the exiv2 tree. The existing terminology ties the include paths to the Solution and the new methodology ties them to the Project and enables them to "float" to any Solution.